Add time picker on readme and example app
