+++ example/pubspec.yaml
@@ -1,6 +1,7 @@
 name: flutter_rounded_date_picker_example
 description: Demonstrates how to use the flutter_rounded_date_picker plugin.
 publish_to: 'none'
+version: 1.0.0+1
 
 environment:
   sdk: ">=2.2.2 <3.0.0"
